+++ backend/src/main/java/com/theladders/api/dto/JobSalaryRequestDeserializer.java
@@ -6,15 +6,17 @@ import tools.jackson.databind.DeserializationContext;
 import tools.jackson.databind.JsonNode;
 import tools.jackson.databind.ValueDeserializer;
 
+import java.math.BigDecimal;
+
 public class JobSalaryRequestDeserializer extends ValueDeserializer<JobSalaryRequest> {
 
     @Override
     public JobSalaryRequest deserialize(JsonParser parser, DeserializationContext context) throws JacksonException {
         JsonNode node = context.readTree(parser);
         if (node.isNumber()) {
-            return new JobSalaryRequest(node.asLong(), "USD", "Annual");
+            return new JobSalaryRequest(node.asDecimal(), "USD", "Annual");
         }
-        Long value = node.hasNonNull("value") ? node.get("value").asLong() : null;
+        BigDecimal value = node.hasNonNull("value") ? node.get("value").asDecimal() : null;
         String currency = node.hasNonNull("currency") ? node.get("currency").asString() : null;
         String unit = node.hasNonNull("unit") ? node.get("unit").asString() : null;
         return new JobSalaryRequest(value, currency, unit);
